+++ src/hooks/useSearchForm.ts
@@ -1,5 +1,7 @@
 import { ChangeEvent, useState } from 'react';
 
+import { validateSearchInput } from '../utils/validationSearchForm';
+
 interface Props {
   onSubmit: (keyword: string) => void;
 }
@@ -9,17 +11,6 @@ const useSearhForm = ({ onSubmit }: Props) => {
   const [error, setError] = useState({
     keyword: '',
   });
-
-  const ERROR_MESSAGE_NO_BLANK = '공백 문자는 허용되지 않습니다.';
-  const ERROR_MESSAGE_REQUIRED_KEYWORD = '검색어를 입력해주세요';
-  const validateSearchInput = (keyword: string) => {
-    const error = { keyword: '' };
-
-    if (keyword.match(/[\s]/g)) error.keyword = ERROR_MESSAGE_NO_BLANK;
-    if (keyword.length === 0) error.keyword = ERROR_MESSAGE_REQUIRED_KEYWORD;
-
-    return error;
-  };
 
   const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
     const keyword = e.target.value;
